Replace SkinCache with WeakMap
