+++ src/renderer/Skin.js
@@ -2,7 +2,6 @@ export default class Skin {
   constructor(renderer) {
     this.renderer = renderer;
     this.gl = renderer.gl;
-    this.used = true;
   }
 
   // Get the skin's texture for a given (screen-space) scale.
